Pequeños cambios en combate

+++ src/personajes/enemigo.cpp
@@ -80,7 +80,7 @@ std::string Enemigo::atacar(Personaje& p) {
     /* Formula de dano */
     int dmg = (getAtaqueFisico() * getAtaqueFisico() / (getAtaqueFisico() +
               p.getDefensaFisica()));
-    dmg = dmg * (bonuses + variacion);
+    dmg = dmg * static_cast<int>(bonuses + variacion);
 
     /* Realizar el ataque */
     p.aumentarVida(-1 * dmg);
@@ -110,7 +110,7 @@ std::string Enemigo::atacar(Personaje& p, int ataqueFisico) {
     /* Formula de dano */
     int dmg = (ataqueFisico * ataqueFisico / (ataqueFisico +
               p.getDefensaFisica()));
-    dmg = dmg * (bonuses + variacion);
+    dmg = dmg * static_cast<int>(bonuses + variacion);
 
     /* Realizar el ataque */
     p.aumentarVida(-1 * dmg);
